refactor: app.js 디렉터리 변경
